Added product module  module

+++ database/migrations/2025_07_11_070301_create_couriers_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('couriers', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('branch_id');
+            $table->unsignedBigInteger('branch_id');
             $table->softDeletes();
             $table->timestamps();
         });
