+++ database/migrations/2025_07_11_085213_create_notifications_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('email');
             $table->text('email_list');
-            $table->integer('branch_id');
+            $table->unsignedBigInteger('branch_id');
             $table->softDeletes();
             $table->timestamps();
         });
